# app/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gener-AI-tions API",
    "description": "API Documentation for the Gener-AI-tions application.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Main Server"
    }
  ],
  "paths": {
...
    "/api/relationships": {
      "post": {
        "summary": "Create Relationship",
        "requestBody": {
          "description": "Relationship data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipCreate"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Relationship created successfully" }
        }
      },
      "delete": {
        "summary": "Delete Relationship",
        "parameters": [
          {
            "name": "relationship_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID of the relationship to delete"
          }
        ],
        "responses": {
          "200": { "description": "Relationship deleted successfully" },
          "404": { "description": "Relationship not found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "RelationshipCreate": {
        "type": "object",
        "properties": {
          "individual_id": { "type": "integer" },
          "related_id": { "type": "integer" },
          "relationship_type": { "type": "string" }
        },
        "required": ["individual_id", "related_id", "relationship_type"]
      }
    }
  }
}